% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{A function that takes in the country's name and
returns a graph that shows the population from 1950 to 2020.}
\usage{
CountryPopulation(countryName)
}
\arguments{
\item{countryName}{the country name as a string}
}
\value{
A graph the shows country's population from 1950 to 2020.
}
\description{
A function that takes in the country's name and
returns a graph that shows the population from 1950 to 2020.
}
