% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{average attendance at each World Cup dependent on the host country.}
\format{
a data frame with observations
\describe{
  \item{WorldCup}{The hosting country and year.}
  \item{Matches}{the number of matches held.}
  \item{Totalattendance}{the total number of attendance.}
  \item{Averageattendance}{the average number of attendance.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
average attendance at each World Cup dependent on the host country.
}
\keyword{datasets}
