% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{United Nations estimated populations for all countries
 that are part of the UN.}
\format{
a data frame with observations
\describe{
  \item{Country Name}{The date of observation as a POSIX date format.}
  \item{1950:2020}{Years of observations.}
}
}
\source{
\url{United Nations}
}
\usage{
WorldPopulation
}
\description{
United Nations estimated populations for all countries
 that are part of the UN.
}
\keyword{datasets}
